Revert "Fixes and checks to adding and inserting players"
This reverts commit f53f7c0c9cd88feb38b95a943bd27d60eacc6786.
