+++ Assets/Art/ImageEffects/PixelationImageEffect.mat
@@ -56,7 +56,7 @@ Material:
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     m_Floats:
-    - _AspectRatio: 1.8752228
+    - _AspectRatio: 1.778
     - _BumpScale: 1
     - _Columns: 128
     - _Cutoff: 0.5
@@ -70,7 +70,7 @@ Material:
     - _OcclusionStrength: 1
     - _OrthoMax: 20
     - _OrthoMin: 5
-    - _OrthoScale: 5
+    - _OrthoScale: 18.769999
     - _Parallax: 0.02
     - _PixelAmmount: 128
     - _PixelAmmountH: 256
