--- Assets/Scripts/Team.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2cca22e1febb0e746969610acd405175
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
